Bump url dependency 0.7.9 -> 0.10.1
When bringing in webpack via npm, and when trying to commit the
node_modules, the following error would happen:

> fatal: Not a git repository:
> /Users/Mathias/Projects/punycode.js/.git/modules/vendor/docdown

Essentially, the
node_modules/webpack/node_modules/node-libs-browser/node_modules/url/node_modules/punycode/vendor/docdown
directory has a .git file in it. It looks like this problem has been
fixed in newer versions. However, node-libs-browser is using an older
version of the url package that has this problem. This commit updates
the version of the url package dependency to resolve this issue.

Fixes #9.